Allow role BUYER to edit field sellerCoOwner in state REVIEW When checking if approved has approved, make sure to check all roles
