+++ libs/application/templates/transport-authority/change-co-owner-of-vehicle/src/utils/hasReviewerApproved.ts
@@ -1,11 +1,38 @@
 import { FormValue } from '@island.is/application/types'
-import { getReviewerInfo } from './getReviewerInfo'
+import { getValueViaPath } from '@island.is/application/core'
+import { CoOwnersInformation, OwnerCoOwnersInformation } from '../shared'
 
 export const hasReviewerApproved = (
   reviewerNationalId: string,
   answers: FormValue,
 ) => {
-  const reviewer = getReviewerInfo(reviewerNationalId, answers)
+  // Check if reviewer is old co-owner and has not approved
+  const ownerCoOwners = getValueViaPath(
+    answers,
+    'ownerCoOwners',
+    [],
+  ) as OwnerCoOwnersInformation[]
+  const ownerCoOwner = ownerCoOwners.find(
+    (x) => x.nationalId === reviewerNationalId,
+  )
+  if (ownerCoOwner) {
+    const hasApproved = ownerCoOwner?.approved || false
+    if (!hasApproved) return false
+  }
 
-  return reviewer ? reviewer.approved || false : true
+  // Check if reviewer is new co-owner and has not approved
+  const coOwners = getValueViaPath(
+    answers,
+    'coOwners',
+    [],
+  ) as CoOwnersInformation[]
+  const coOwner = coOwners
+    .filter(({ wasRemoved }) => wasRemoved !== 'true')
+    .find((x) => x.nationalId === reviewerNationalId)
+  if (coOwner) {
+    const hasApproved = coOwner?.approved || false
+    if (!hasApproved) return false
+  }
+
+  return true
 }
