+++ libs/application/templates/transport-authority/change-operator-of-vehicle/src/utils/hasReviewerApproved.ts
@@ -1,11 +1,38 @@
 import { FormValue } from '@island.is/application/types'
-import { getReviewerInfo } from './getReviewerInfo'
+import { getValueViaPath } from '@island.is/application/core'
+import { OperatorInformation, UserInformation } from '../shared'
 
 export const hasReviewerApproved = (
   reviewerNationalId: string,
   answers: FormValue,
 ) => {
-  const reviewer = getReviewerInfo(reviewerNationalId, answers)
+  // Check if reviewer is owner coowner and has not approved
+  const ownerCoOwners = getValueViaPath(
+    answers,
+    'ownerCoOwner',
+    [],
+  ) as UserInformation[]
+  const ownerCoOwner = ownerCoOwners.find(
+    (ownerCoOwner) => ownerCoOwner.nationalId === reviewerNationalId,
+  )
+  if (ownerCoOwner) {
+    const hasApproved = ownerCoOwner?.approved || false
+    if (!hasApproved) return false
+  }
 
-  return reviewer ? reviewer.approved || false : true
+  // Check if reviewer is operator and has not approved
+  const operators = getValueViaPath(
+    answers,
+    'operators',
+    [],
+  ) as OperatorInformation[]
+  const operator = operators
+    .filter(({ wasRemoved }) => wasRemoved !== 'true')
+    .find((operator) => operator.nationalId === reviewerNationalId)
+  if (operator) {
+    const hasApproved = operator?.approved || false
+    if (!hasApproved) return false
+  }
+
+  return true
 }
