+++ libs/application/templates/transport-authority/transfer-of-vehicle-ownership/src/lib/TransferOfVehicleOwnershipTemplate.ts
@@ -193,10 +193,10 @@ const template: ApplicationTemplate<
               write: {
                 answers: [
                   'sellerCoOwner',
-                  'buyerCoOwnerAndOperator',
-                  'rejecter',
-                  'insurance',
                   'buyer',
+                  'buyerCoOwnerAndOperator',
+                  'insurance',
+                  'rejecter',
                 ],
               },
               read: 'all',
@@ -210,11 +210,12 @@ const template: ApplicationTemplate<
                 ),
               write: {
                 answers: [
-                  'buyerCoOwnerAndOperator',
-                  'insurance',
+                  'sellerCoOwner',
                   'buyer',
-                  'rejecter',
+                  'buyerCoOwnerAndOperator',
                   'buyerMainOperator',
+                  'insurance',
+                  'rejecter',
                 ],
               },
               read: 'all',
