+++ libs/application/templates/transport-authority/transfer-of-vehicle-ownership/src/utils/getApproveAnswers.ts
@@ -79,7 +79,6 @@ export const getApproveAnswers = (
           name: coOwner.name,
           email: coOwner.email,
           phone: coOwner.phone,
-          type: coOwner.type,
           approved:
             coOwner.nationalId === reviewerNationalId
               ? true
