+++ libs/application/templates/transport-authority/transfer-of-vehicle-ownership/src/utils/hasReviewerApproved.ts
@@ -1,11 +1,51 @@
 import { FormValue } from '@island.is/application/types'
-import { getReviewerInfo } from './getReviewerInfo'
+import { getValueViaPath } from '@island.is/application/core'
+import { CoOwnerAndOperator, UserInformation } from '../shared'
 
 export const hasReviewerApproved = (
   reviewerNationalId: string,
   answers: FormValue,
 ) => {
-  const reviewer = getReviewerInfo(reviewerNationalId, answers)
+  // Check if reviewer is buyer and has not approved
+  if (
+    (getValueViaPath(answers, 'buyer.nationalId', '') as string) ===
+    reviewerNationalId
+  ) {
+    const buyer = getValueViaPath(answers, 'buyer') as UserInformation
+    const hasApproved = buyer?.approved || false
+    if (!hasApproved) return false
+  }
 
-  return reviewer ? reviewer.approved || false : true
+  // Check if reviewer is buyers coowner or operator and has not approved
+  const buyerCoOwnersAndOperators = getValueViaPath(
+    answers,
+    'buyerCoOwnerAndOperator',
+    [],
+  ) as CoOwnerAndOperator[]
+  const buyerCoOwnerAndOperator = buyerCoOwnersAndOperators
+    .filter(({ wasRemoved }) => wasRemoved !== 'true')
+    .find(
+      (coOwnerOrOperator) =>
+        coOwnerOrOperator.nationalId === reviewerNationalId,
+    )
+  if (buyerCoOwnerAndOperator) {
+    const hasApproved = buyerCoOwnerAndOperator?.approved || false
+    if (!hasApproved) return false
+  }
+
+  // Check if reviewer is sellers coowner and has not approved
+  const sellerCoOwners = getValueViaPath(
+    answers,
+    'sellerCoOwner',
+    [],
+  ) as CoOwnerAndOperator[]
+  const sellerCoOwner = sellerCoOwners.find(
+    (coOwner) => coOwner.nationalId === reviewerNationalId,
+  )
+  if (sellerCoOwner) {
+    const hasApproved = sellerCoOwner?.approved || false
+    if (!hasApproved) return false
+  }
+
+  return true
 }
