+++ libs/application/templates/transport-authority/transfer-of-vehicle-ownership/src/utils/index.ts
@@ -8,7 +8,6 @@ export const formatPhoneNumber = (value: string): string =>
 
 export { getSelectedVehicle } from './getSelectedVehicle'
 export { getReviewSteps } from './getReviewSteps'
-export { getReviewerInfo } from './getReviewerInfo'
 export { hasReviewerApproved } from './hasReviewerApproved'
 export { getApproveAnswers } from './getApproveAnswers'
 export { isLastReviewer } from './isLastReviewer'
